2018-8-6

+++ src/main/java/com/ssm/adminlte/entity/Permission.java
@@ -1,7 +1,7 @@
 package com.ssm.adminlte.entity;
 
 public class Permission {
-    private Long id;
+    private Integer id;
 
     private String name;
 
@@ -19,15 +19,15 @@ public class Permission {
 
     private String available;
 
-    public Long getId() {
-        return id;
-    }
+    public Integer getId() {
+		return id;
+	}
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+	public void setId(Integer id) {
+		this.id = id;
+	}
 
-    public String getName() {
+	public String getName() {
         return name;
     }
